map a cube map to a cube
