+++ assets/cases/copyBuffersToTexture.meta
@@ -1,0 +1,12 @@
+{
+  "ver": "1.1.0",
+  "importer": "directory",
+  "imported": true,
+  "uuid": "fe59f68c-4b81-43fe-844a-33d993e6cd1e",
+  "files": [],
+  "subMetas": {},
+  "userData": {
+    "compressionType": {},
+    "isRemoteBundle": {}
+  }
+}
