+++ assets/cases/copyBuffersToTexture/copyBuffersToTexture.scene.meta
@@ -1,0 +1,11 @@
+{
+  "ver": "1.1.37",
+  "importer": "scene",
+  "imported": true,
+  "uuid": "7cdb5198-f6da-42f6-97ba-8ffc94a45626",
+  "files": [
+    ".json"
+  ],
+  "subMetas": {},
+  "userData": {}
+}
