+++ assets/cases/copyBuffersToTexture/copyBuffersToTexture.ts.meta
@@ -1,0 +1,9 @@
+{
+  "ver": "4.0.23",
+  "importer": "typescript",
+  "imported": true,
+  "uuid": "ad4eeacc-4108-44f8-bffb-b50809e4844b",
+  "files": [],
+  "subMetas": {},
+  "userData": {}
+}
