+++ assets/cases/copyBuffersToTexture/cubeMap.effect
@@ -5,10 +5,18 @@ CCEffect %{
       frag: complex-fs:frag
       priority: max - 10
       rasterizerState:
-        cullMode: none
+        cullMode: back
+      embeddedMacros: { USE_RGBE_CUBEMAP: true }
       depthStencilState:
         depthTest: true
         depthWrite: false
+      blendState:
+        targets:
+        - blend: true
+          blendSrc: src_alpha
+          blendDst: one_minus_src_alpha
+          blendDstAlpha: one_minus_src_alpha
+
 }%
 
 CCProgram complex-vs %{
@@ -16,12 +24,11 @@ CCProgram complex-vs %{
   #include <cc-global>
   #include <decode>
 
-  in vec3 v_position;
   out mediump vec4 viewDir;
 
   vec4 vert () {
-    viewDir = vec4(v_position, 1.0);
-    return viewDir;
+    CCDecode(viewDir);
+    return cc_matViewProj * viewDir;
   }
 }%
 
@@ -35,11 +42,13 @@ CCProgram complex-fs %{
 
   in mediump vec4 viewDir;
 
+  uniform samplerCube cubeMap;
+
   vec4 frag () {
     #if USE_RGBE_CUBEMAP
-      vec3 c = unpackRGBE(texture(cc_environment, viewDir.xyz));
+      vec3 c = unpackRGBE(texture(cubeMap, viewDir.xyz));
     #else
-      vec3 c = SRGBToLinear(texture(cc_environment, viewDir.xyz).rgb);
+      vec3 c = SRGBToLinear(texture(cubeMap, viewDir.xyz).rgb);
     #endif
     return CCFragOutput(vec4(c * cc_ambientSky.w, 1.0));
   }
