+++ assets/cases/copyBuffersToTexture/cubeMap.effect.meta
@@ -1,0 +1,11 @@
+{
+  "ver": "1.4.9",
+  "importer": "effect",
+  "imported": true,
+  "uuid": "ecba09f9-9a73-4449-8bb3-760b27e14bda",
+  "files": [
+    ".json"
+  ],
+  "subMetas": {},
+  "userData": {}
+}
