+++ assets/cases/copyBuffersToTexture/cubeMap.ts.meta
@@ -1,0 +1,9 @@
+{
+  "ver": "4.0.23",
+  "importer": "typescript",
+  "imported": true,
+  "uuid": "a46f9c7d-b892-4b9f-a277-d98f062186e4",
+  "files": [],
+  "subMetas": {},
+  "userData": {}
+}
