+++ assets/cases/copyBuffersToTexture/cubeMapMat.mtl
@@ -4,14 +4,14 @@
   "_objFlags": 0,
   "_native": "",
   "_effectAsset": {
-    "__uuid__": "1baf0fc9-befa-459c-8bdd-af1a450a0319",
+    "__uuid__": "ecba09f9-9a73-4449-8bb3-760b27e14bda",
     "__expectedType__": "cc.EffectAsset"
   },
   "_techIdx": "0",
   "_defines": [
-    {},
-    {},
-    {}
+    {
+      "USE_RGBE_CUBEMAP": true
+    }
   ],
   "_states": [
     {
@@ -22,24 +22,9 @@
           {}
         ]
       }
-    },
-    {},
-    {}
+    }
   ],
   "_props": [
-    {
-      "roughness": 0.769,
-      "metallic": 1,
-      "SpecularIntensity": 1,
-      "emissive": {
-        "__type__": "cc.Color",
-        "r": 255,
-        "g": 255,
-        "b": 255,
-        "a": 0
-      }
-    },
-    {},
     {}
   ]
 }
