+++ assets/cases/copyBuffersToTexture/cubeMapMat.mtl.meta
@@ -1,0 +1,11 @@
+{
+  "ver": "1.0.12",
+  "importer": "material",
+  "imported": true,
+  "uuid": "2c2bc0f5-62bd-44c2-8d02-aaa137fc1e94",
+  "files": [
+    ".json"
+  ],
+  "subMetas": {},
+  "userData": {}
+}
