+++ assets/cases/copyBuffersToTexture/planarMat.mtl
@@ -7,7 +7,7 @@
     "__uuid__": "1baf0fc9-befa-459c-8bdd-af1a450a0319",
     "__expectedType__": "cc.EffectAsset"
   },
-  "_techIdx": "0",
+  "_techIdx": "1",
   "_defines": [
     {
       "USE_ALBEDO_MAP": true
@@ -31,7 +31,7 @@
   "_props": [
     {
       "mainTexture": {
-        "__uuid__": "23f89a10-e703-433a-8546-ab236267968b@6c48a",
+        "__uuid__": "42d2b4a0-61d5-4938-876b-477a833ed0c4@6c48a",
         "__expectedType__": "cc.Texture2D"
       }
     },
