+++ assets/cases/copyBuffersToTexture/planarMat.mtl.meta
@@ -1,0 +1,1 @@
+{"ver":"1.0.12","importer":"material","imported":true,"uuid":"819f168a-55c6-4b62-aa9b-e6af4d19078f","files":[".json"],"subMetas":{},"userData":{}}
